refactor:: DayType 파일 utils -> common/types로 이동
