+++ src/common/types/day-type.utils.ts
