+++ src/schedule/dto/create-schedule-request.dto.ts
@@ -6,8 +6,8 @@ import {
   IsOptional,
   IsString,
 } from 'class-validator';
+import { DayType } from 'src/common/types/day-type.utils';
 import { IsTime } from 'src/decorators/time.decorator';
-import { DayType } from 'src/utils/day-type.utils';
 
 export class CreateScheduleRequestDto {
   @ApiProperty({ description: '시간표 ID' })
